cleanup cg code correct prototype for cg segment identifier, now only segment_id prototype is used use segment_id prototype for C, C++ and F77 compilers unify symbol names for segment identifier variables use short type for segment identifier variables globaly add new header file objout.h to share prototypes related to object file

+++ bld/cg/h/cg.h
@@ -86,7 +86,7 @@ typedef void *          patch_handle;   /*  used to stuff patchable ints into tr
 
 typedef void            (*cg_callback)( callback_handle );
 
-typedef int             segment_id;     /*  parm to BESetSeg() */
+typedef short           segment_id;     /*  must be signed */
 
 typedef unsigned_32     dbg_type;       /*  symbolic debugging type handle */
 typedef void            *dbg_struct;
